requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
anyio==3.6.2
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
coverage==6.4.1
cycler==0.11.0
Cython==0.29.30
DateTime==4.4
distlib==0.3.4
docopt==0.6.2
docutils==0.18.1
filelock==3.7.0
Flask==1.1.4
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
green==3.4.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.3
iso8601==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
Js2Py==0.71
keyring==21.1.1
kiwisolver==1.4.2
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pipwin==0.5.2
platformdirs==2.5.2
py==1.8.2
# PyAudio @ file:///C:/Users/xiaox/pipwin/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
pygame==2.1.2
pyjsparser==2.7.1
pynput==1.7.6
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
requests==2.27.1
rfc3986==1.5.0
simple-chalk==0.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
tornado==6.0.4
twitchio @ git+http://github.com/TwitchIO/TwitchIO@7a46bd17d4c3f1eac53d9481c8c43024d3a7ae6a
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
Unidecode==1.3.4
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==1.0.1
xkcd-wrapper==1.0.1
yarl==1.7.2
zope.interface==5.4.0

twitchio~=2.2.0
geopy~=2.3.0
